# public/swagger/v1/swagger.json
{
	"components": {
...
		"schemas": {
			"Link": {
				"properties": {
					"rel": {
						"type": "string"
					},
					"href": {
						"type": "string"
					}
				},
				"required": [
					"rel",
					"href"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Authority": {
				"properties": {
					"localAuthorityId": {
						"type": "number",
						"format": "double"
					},
					"localAuthorityIdCode": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"friendlyName": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"schemeURL": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"regionName": {
						"type": "string"
					},
					"fileName": {
						"type": "string"
					},
					"fileNameWelsh": {
						"type": "string"
					},
					"establishmentCount": {
						"type": "number",
						"format": "double"
					},
					"creationDate": {
						"type": "string",
						"format": "date-time"
					},
					"lastPublishedDate": {
						"type": "string",
						"format": "date-time"
					},
					"schemeType": {
						"type": "number",
						"format": "double"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"localAuthorityId",
					"localAuthorityIdCode",
					"name",
					"friendlyName",
					"url",
					"schemeURL",
					"email",
					"regionName",
					"fileName",
					"fileNameWelsh",
					"establishmentCount",
					"creationDate",
					"lastPublishedDate",
					"schemeType",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Meta": {
				"properties": {
					"publisher": {
						"type": "string"
					},
					"licence": {
						"type": "string"
					},
					"documentation": {
						"type": "string"
					},
					"version": {
						"type": "string"
					},
					"comment": {
						"type": "string"
					},
					"hasFormat": {
						"items": {
							"type": "string"
						},
						"type": "array"
					},
					"limit": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"publisher",
					"licence",
					"documentation",
					"version",
					"comment",
					"hasFormat",
					"limit"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAuthorityDetailPaginationGetResponse": {
				"properties": {
					"authorities": {
						"items": {
							"$ref": "#/components/schemas/Authority"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"authorities",
					"meta",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"AuthorityBasic": {
				"properties": {
					"localAuthorityId": {
						"type": "number",
						"format": "double"
					},
					"localAuthorityIdCode": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"establishmentCount": {
						"type": "number",
						"format": "double"
					},
					"schemeType": {
						"type": "number",
						"format": "double"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"localAuthorityId",
					"localAuthorityIdCode",
					"name",
					"establishmentCount",
					"schemeType",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IAuthorityDetailBasicPaginationGetResponse": {
				"properties": {
					"authorities": {
						"items": {
							"$ref": "#/components/schemas/AuthorityBasic"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"authorities",
					"meta",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IAuthorityDetailByIDGetResponse": {
				"properties": {
					"localAuthorityId": {
						"type": "number",
						"format": "double"
					},
					"localAuthorityIdCode": {
						"type": "string"
					},
					"name": {
						"type": "string"
					},
					"friendlyName": {
						"type": "string"
					},
					"url": {
						"type": "string"
					},
					"schemeURL": {
						"type": "string"
					},
					"email": {
						"type": "string"
					},
					"regionName": {
						"type": "string"
					},
					"fileName": {
						"type": "string"
					},
					"fileNameWelsh": {
						"type": "string"
					},
					"establishmentCount": {
						"type": "number",
						"format": "double"
					},
					"creationDate": {
						"type": "string",
						"format": "date-time"
					},
					"lastPublishedDate": {
						"type": "string",
						"format": "date-time"
					},
					"schemeType": {
						"type": "number",
						"format": "double"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"localAuthorityId",
					"localAuthorityIdCode",
					"name",
					"friendlyName",
					"url",
					"schemeURL",
					"email",
					"regionName",
					"fileName",
					"fileNameWelsh",
					"establishmentCount",
					"creationDate",
					"lastPublishedDate",
					"schemeType",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"BusinessType": {
				"properties": {
					"businessTypeID": {
						"type": "number",
						"format": "double"
					},
					"businessTypeName": {
						"type": "string"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"businessTypeID",
					"businessTypeName",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBusinessTypesDetailPaginationGetResponse": {
				"properties": {
					"businessTypes": {
						"items": {
							"$ref": "#/components/schemas/BusinessType"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"businessTypes",
					"meta",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"IBusinessTypesDetailBasicPaginationGetResponse": {
				"properties": {
					"businessTypes": {
						"items": {
							"$ref": "#/components/schemas/BusinessType"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"businessTypes",
					"meta",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IBusinessTypesDetailByIdGetResponse": {
				"properties": {
					"businessTypeID": {
						"type": "number",
						"format": "double"
					},
					"businessTypeName": {
						"type": "string"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"businessTypeID",
					"businessTypeName",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"Country": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"nameKey": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"nameKey",
					"code",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICountryDetailPaginationGetResponse": {
				"properties": {
					"countries": {
						"items": {
							"$ref": "#/components/schemas/Country"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"countries",
					"meta",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ICountryDetailBasicPaginationGetResponse": {
				"properties": {
					"countries": {
						"items": {
							"$ref": "#/components/schemas/Country"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/Meta"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"countries",
					"meta",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"ICountryDetailByIdGetResponse": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"nameKey": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"nameKey",
					"code",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Region": {
				"properties": {
					"id": {
						"type": "number",
						"format": "double"
					},
					"name": {
						"type": "string"
					},
					"nameKey": {
						"type": "string"
					},
					"code": {
						"type": "string"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"id",
					"name",
					"nameKey",
					"code",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
			"MetaRegion": {
				"properties": {
					"dataSource": {
						"type": "string"
					},
					"extractDate": {
						"type": "string",
						"format": "date-time"
					},
					"itemCount": {
						"type": "number",
						"format": "double"
					},
					"returncode": {
						"type": "string"
					},
					"totalCount": {
						"type": "number",
						"format": "double"
					},
					"totalPages": {
						"type": "number",
						"format": "double"
					},
					"pageSize": {
						"type": "number",
						"format": "double"
					},
					"pageNumber": {
						"type": "number",
						"format": "double"
					}
				},
				"required": [
					"dataSource",
					"extractDate",
					"itemCount",
					"returncode",
					"totalCount",
					"totalPages",
					"pageSize",
					"pageNumber"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"IRegionDetailBasicPaginationGetResponse": {
				"properties": {
					"regions": {
						"items": {
							"$ref": "#/components/schemas/Region"
						},
						"type": "array"
					},
					"meta": {
						"$ref": "#/components/schemas/MetaRegion"
					},
					"links": {
						"items": {
							"$ref": "#/components/schemas/Link"
						},
						"type": "array"
					}
				},
				"required": [
					"regions",
					"meta",
					"links"
				],
				"type": "object",
				"additionalProperties": false
			},
...
		},
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"name": "authorization",
				"in": "header",
				"bearerFormat": "JWT"
			},
			"apiKeyAuth": {
				"type": "apiKey",
				"in": "header",
				"name": "X-API-KEY"
			}
		}
	},
	"info": {
		"title": "Food Standards Agency API",
		"version": "v1",
		"description": "The FSA Food Alerts API provides access to current and recent Food Alerts: Allergy Alerts (AA), Product Recall Information Notices (PRIN) and Food Alerts for Action (FAFA). It provides applications with the facility to list alerts matching some filter criterion, and to retrieve a description of an alert.",
		"contact": {
			"name": "Oraclo LTD",
			"email": "admin@orcalo.co.uk",
			"url": "https://orcalo.co.uk/"
		}
	},
	"openapi": "3.0.0",
	"paths": {
		"/v1/authorities/{pageNumber}/{pageSize}": {
			"get": {
				"operationId": "GetAuthorityDetailPagination",
				"responses": {
					"200": {
						"description": "IAuthorityDetailPaginationGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAuthorityDetailPaginationGetResponse"
								}
							}
						}
					}
				},
				"summary": "Returns details of all authorities, page parameters allow for page number and size specification.",
				"tags": [
					"Authorities"
				],
				"security": [],
				"parameters": [
					{
						"description": "Must be Greater Than <b>0</b>",
						"in": "path",
						"name": "pageNumber",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Must be Greater Than <b>0</b>",
						"in": "path",
						"name": "pageSize",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/authorities/basic/{pageNumber}/{pageSize}": {
			"get": {
				"operationId": "GetRegionDetailBasicPagination",
				"responses": {
					"200": {
						"description": "IAuthorityDetailBasicPaginationGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAuthorityDetailBasicPaginationGetResponse"
								}
							}
						}
					}
				},
				"summary": "Returns a basic list of authorities (id/name), details can be retrieved by calling authorities/{id}, page parameters allow for page number and size specification.",
				"tags": [
					"Authorities"
				],
				"security": [],
				"parameters": [
					{
						"description": "Must be greater than <b>0</b>",
						"in": "path",
						"name": "pageNumber",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Must be greater than <b>0</b>",
						"in": "path",
						"name": "pageSize",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/authorities/{id}": {
			"get": {
				"operationId": "GetAuthorityDetailById",
				"responses": {
					"200": {
						"description": "IAuthorityDetailByIDGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IAuthorityDetailByIDGetResponse"
								}
							}
						}
					}
				},
				"summary": "Returns details of a single authority, selected by Id.",
				"tags": [
					"Authorities"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/businesstypes/{pageNumber}/{pageSize}": {
			"get": {
				"operationId": "GetBusinessTypesDetailPagination",
				"responses": {
					"200": {
						"description": "IBusinessTypesDetailPaginationGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBusinessTypesDetailPaginationGetResponse"
								}
							}
						}
					}
				},
				"summary": "Returns details of all business types, page parameters allow for page number and size specification.",
				"tags": [
					"BusinessTypes"
				],
				"security": [],
				"parameters": [
					{
						"description": "Must be Greater Than <b>0</b>",
						"in": "path",
						"name": "pageNumber",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Must be Greater Than <b>0</b>",
						"in": "path",
						"name": "pageSize",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/businesstypes/basic/{pageNumber}/{pageSize}": {
			"get": {
				"operationId": "GetBusinessTypesDetailBasicPagination",
				"responses": {
					"200": {
						"description": "IBusinessTypesDetailBasicPaginationGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBusinessTypesDetailBasicPaginationGetResponse"
								}
							}
						}
					}
				},
				"summary": "Returns a basic list of business types, details can be retrieved by calling countries/{id}, page parameters allow for page number and size specification.",
				"tags": [
					"BusinessTypes"
				],
				"security": [],
				"parameters": [
					{
						"description": "Must be greater than <b>0</b>",
						"in": "path",
						"name": "pageNumber",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Must be greater than <b>0</b>",
						"in": "path",
						"name": "pageSize",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/businesstypes/{id}": {
			"get": {
				"operationId": "GetBusinessTypesDetailById",
				"responses": {
					"200": {
						"description": "IBusinessTypesDetailByIdGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IBusinessTypesDetailByIdGetResponse"
								}
							}
						}
					}
				},
				"summary": "Returns details of a single business type, selected by Id.",
				"tags": [
					"BusinessTypes"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
		"/v1/countries/{pageNumber}/{pageSize}": {
			"get": {
				"operationId": "GetCountryDetailPagination",
				"responses": {
					"200": {
						"description": "ICountryDetailPaginationGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICountryDetailPaginationGetResponse"
								}
							}
						}
					}
				},
				"summary": "Returns details of all countries, page parameters allow for page number and size specification.",
				"tags": [
					"Countries"
				],
				"security": [],
				"parameters": [
					{
						"description": "Must be greater than <b>0</b>",
						"in": "path",
						"name": "pageNumber",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Must be greater than <b>0</b>",
						"in": "path",
						"name": "pageSize",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/v1/countries/basic/{pageNumber}/{pageSize}": {
			"get": {
				"operationId": "GetCountryDetailBasicPagination",
				"responses": {
					"200": {
						"description": "ICountryDetailBasicPaginationGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICountryDetailBasicPaginationGetResponse"
								}
							}
						}
					}
				},
				"summary": "Returns a basic list of countries, details can be retrieved by calling countries/{id}, page parameters allow for page number and size specification.",
				"tags": [
					"Countries"
				],
				"security": [],
				"parameters": [
					{
						"description": "Must be greater than <b>0</b>",
						"in": "path",
						"name": "pageNumber",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Must be greater than <b>0</b>",
						"in": "path",
						"name": "pageSize",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
		"/v1/countries/{id}": {
			"get": {
				"operationId": "GetCountryDetailById",
				"responses": {
					"200": {
						"description": "ICountryDetailByIdGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ICountryDetailByIdGetResponse"
								}
							}
						}
					}
				},
				"summary": "Returns details of a single country, selected by Id.",
				"tags": [
					"Countries"
				],
				"security": [],
				"parameters": [
					{
						"description": "Must be greater than <b>0</b>",
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"format": "double",
							"type": "number"
						}
					}
				]
			}
		},
...
		"/v1/regions/basic/{pageNumber}/{pageSize}": {
			"get": {
				"operationId": "GetRegionDetailBasicPagination",
				"responses": {
					"200": {
						"description": "IRegionDetailBasicPaginationGetResponse",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/IRegionDetailBasicPaginationGetResponse"
								}
							}
						}
					}
				},
				"summary": "Returns a basic list of regions, details can be retrieved by calling regions{id}, page parameters allow for page number and size specification.",
				"tags": [
					"Regions"
				],
				"security": [],
				"parameters": [
					{
						"description": "Must be greater than <b>0</b>",
						"in": "path",
						"name": "pageNumber",
						"required": true,
						"schema": {
							"type": "string"
						}
					},
					{
						"description": "Must be greater than <b>0</b>",
						"in": "path",
						"name": "pageSize",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}